+++ exoline/exo.py
@@ -74,6 +74,7 @@ import requests
 import yaml
 import importlib
 import humanize
+import blessings
 
 from pyonep import onep
 from pyonep import provision
@@ -97,6 +98,8 @@ DEFAULT_CONFIG = '~/.exoline'
 SCRIPT_LIMIT_BYTES = 16 * 1024
 
 PERF_DATA = []
+
+colored_terminal = blessings.Terminal()
 
 cmd_doc = OrderedDict([
     ('read',
@@ -744,7 +747,20 @@ class ExoRPC():
     regex_tweeid = re.compile("rid\.[0-9a-fA-F]{5}")
 
     class RPCException(Exception):
-        pass
+        def __init__(self, *args):
+            try:
+                err, conditions = args[0].split(" ", 1)
+                if err == "invalid":
+                    url = "https://pyonep.readthedocs.org/en/latest/errors/invalid.html"
+                elif err == "auth":
+                    url = "https://pyonep.readthedocs.org/en/latest/errors/auth.html"
+                else:
+                    url = "https://pyonep.readthedocs.org/en/latest/errors/general.html"
+                self.message = "Error: %s\n\tArguments: %s\n\tFor more information, visit: %s"%(err, conditions, url)
+            except:
+                self.message = args[0]
+        def __str__(self):
+            return self.message
 
     def __init__(self,
                  host=DEFAULT_HOST,
@@ -1267,7 +1283,7 @@ class ExoRPC():
         return response
 
     def listing(self, cik, types, options={}, rid=None):
-        isok, response = self.exo.listing(cik, types, options=options, rid=rid)
+        isok, response = self.exo.listing(cik, types, options=options, resource=rid)
         self._raise_for_response(isok, response)
         return response
 
@@ -1592,37 +1608,53 @@ class ExoRPC():
 
         if twee:
             # colors, of course
-            class bcolors:
-                SPACER = '' if cli_args['--nocolor'] else '\033[0m'
-                NAME = '' if cli_args['--nocolor'] else '\033[0m'
-                TYPE = '' if cli_args['--nocolor'] else '\033[35m'
-                ID = '' if cli_args['--nocolor'] else '\033[32m'
-                VALUE = '' if cli_args['--nocolor'] else '\033[33m'
-                TIMESTAMP = '' if cli_args['--nocolor'] else '\033[34m'
-                PINK = '' if cli_args['--nocolor'] else '\033[35m'
-                MODEL = '' if cli_args['--nocolor'] else '\033[36m'
-                ENDC = '' if cli_args['--nocolor'] else '\033[0m'
+            default = colored_terminal.normal
+
+            if cli_args['--nocolor']:
+                SPACER = default
+                NAME = default
+                TYPE = default
+                ID = default
+                VALUE = default
+                TIMESTAMP = default
+                PINK = default
+                MODEL = default
+                ENDC = default
+            else:
+                SPACER = default
+                NAME = default
+                TYPE = colored_terminal.magenta
+                ID = colored_terminal.green
+                TIMESTAMP = colored_terminal.blue
+                VALUE = colored_terminal.yellow
+                PINK = colored_terminal.magenta
+                MODEL = colored_terminal.cyan
+                ENDC = colored_terminal.normal
 
             # the goal here is to make the line short to provide more room for the value
             # so if there's an alias, just use that, since it's
             # if no alias, then the first ten of the RID and the name
             # if multiple alias, then the first alias
+
             if typ == 'client':
                 if cli_args['--rids']:
-                    tweeid = bcolors.SPACER + 'rid: ' + bcolors.ID + rid
+                    tweeid = SPACER + 'rid: ' + ID + rid
                 else:
-                    tweeid = bcolors.SPACER + 'cik: ' + bcolors.ID + id[5:]
+                    tweeid = SPACER + 'cik: ' + ID + id[5:]
             else:
                 if cli_args['--rids']:
-                    tweeid = bcolors.SPACER + 'rid: ' + bcolors.ID + rid
+                    tweeid = SPACER + 'rid: ' + ID + rid
                 else:
                     if aliases is not None and len(aliases) > 0:
                         tweeid = aliases[0]
                     else:
                         tweeid = 'rid.' + rid[:5]
 
-            displayname = ((name + bcolors.SPACER + ' ') if len(name) > 0 else ' ')
+            displayname = ((name + SPACER + ' ') if len(name) > 0 else ' ')
+
             displaytype = {'dataport': 'dp', 'client': 'cl', 'datarule': 'dr', 'dispatch': 'ds'}[typ]
+            if not cli_args['--nocolor'] and displaytype == "cl":
+                NAME = colored_terminal.underline
             if 'format' in info['description']:
                 displaytype += '.' + {'binary': 'b', 'string': 's', 'float': 'f', 'integer': 'i'}[info['description']['format']]
             else:
@@ -1645,25 +1677,19 @@ class ExoRPC():
                     val = val[:allowed_size] + "..."
 
             self._print_tree_line(
-                bcolors.SPACER +
-                spacer +
-                bcolors.NAME +
-                displayname +
-                ' ' * (maxlen['name'] + 1 - len(name)) +
-                bcolors.TYPE +
-                displaytype + ' ' +
-                bcolors.ID +
-                tweeid +
-                bcolors.SPACER +
-                (' (share)' if 'listing_option' in info and info['listing_option'] == 'activated' else '') +
+                SPACER + spacer +
+                NAME + displayname +
+                ' '*(maxlen['name']-len(name)) +
+                TYPE + displaytype +
+                ' ' +
+                ID + tweeid +
+                SPACER + (' (share)' if 'listing_option' in info and info['listing_option'] == 'activated' else '') +
                 ('' if typ == 'client' else ': ') +
-                bcolors.VALUE +
-                ('' if val is None else val) +
-                bcolors.TIMESTAMP +
-                ('' if timestamp is None or len(timestamp) == 0 else ' (' + timestamp + ')') +
-                bcolors.MODEL +
-                displaymodel +
-                bcolors.ENDC)
+                VALUE + ('' if val is None else val) +
+                TIMESTAMP + ('' if timestamp is None or len(timestamp) == 0 else ' (' + timestamp + ')') +
+                MODEL + displaymodel +
+                ENDC
+                )
         else:
             # standard tree
             if 'format' in info['description']:
@@ -1811,7 +1837,7 @@ probably not valid.".format(cik))
             cik,
             types=['client', 'dataport', 'datarule', 'dispatch'],
             options={},
-            rid={'alias': ''})
+            resource={'alias': ''})
         self._raise_for_response(isok, listing)
         rids = itertools.chain(*[listing[t] for t in listing.keys()])
         self._exomult(cik, [['drop', rid] for rid in rids])
@@ -1989,14 +2015,22 @@ probably not valid.".format(cik))
                     last_status = ''
                     uploaded = False
                     nocolor = platform.system() == 'Windows'
+                    def ifcolor(c):
+                        return colored_terminal.normal if nocolor else c
                     class colors:
-                        PINK = '' if nocolor else '\033[35m'
-                        CYAN = '' if nocolor else '\033[36m'
-                        YELLOW = '' if nocolor else '\033[33m'
-                        GREEN = '' if nocolor else '\033[32m'
-                        RED = '' if nocolor else '\033[31m'
-                        GRAY = '' if nocolor else '\033[1;30m'
-                        ENDC = '' if nocolor else '\033[0m'
+                        SPACER = ifcolor(colored_terminal.normal)
+                        NAME = ifcolor(colored_terminal.normal)
+                        TYPE = ifcolor(colored_terminal.magenta)
+                        ID = ifcolor(colored_terminal.green)
+                        TIMESTAMP = ifcolor(colored_terminal.blue)
+                        VALUE = ifcolor(colored_terminal.yellow)
+                        PINK = ifcolor(colored_terminal.magenta)
+                        MODEL = ifcolor(colored_terminal.cyan)
+                        ENDC = ifcolor(colored_terminal.normal)
+                        GRAY = ifcolor(colored_terminal.gray)
+                        GREEN = ifcolor(colored_terminal.green)
+                        RED = ifcolor(colored_terminal.red)
+
                     def status_color(status):
                         return colors.RED if status == 'error' else colors.GREEN
                     # loop forever
@@ -2083,7 +2117,7 @@ probably not valid.".format(cik))
                     upl()
 
     def lookup_rid(self, cik, cik_to_find):
-        isok, listing = self.exo.listing(cik, types=['client'], options={}, rid={'alias': ''})
+        isok, listing = self.exo.listing(cik, types=['client'], options={}, resource={'alias': ''})
         self._raise_for_response(isok, listing)
 
         for rid in listing['client']:
@@ -2641,8 +2675,8 @@ class ExoUtilities():
     def format_time(cls, sec):
         '''Formats a time interval for human consumption'''
         intervals = [[60 * 60 * 24, 'd'],
-                        [60 * 60, 'h'],
-                        [60, 'm']]
+                     [60 * 60, 'h'],
+                     [60, 'm']]
         text = ""
         for s, label in intervals:
             if sec >= s and sec // s > 0:
@@ -2654,11 +2688,12 @@ class ExoUtilities():
 
     @classmethod
     def handleSystemExit(cls, ex):
-        # Handle SystemExit per https://docs.python.org/2/library/exceptions.html#exceptions.SystemExit
+        # Handle SystemExit per
+        # https://docs.python.org/2/library/exceptions.html#exceptions.SystemExit
         if ex.code is None:
             return 0
         elif isinstance(ex.code, six.string_types):
-            sys.stderr.write(ex.code)
+            sys.stderr.write(ex.code + '\n')
             return 1
         elif type(ex.code is int):
             return ex.code